fix: refactor cert backend selection to avoid circular imports
