+++ src/rasenmaeher_api/db/people.py
@@ -23,7 +23,7 @@ from libadvian.tasks import TaskMaster
 from .base import ORMBaseModel, utcnow
 from ..web.api.middleware.datatypes import MTLSorJWTPayload
 from .errors import NotFound, Deleted, BackendError, CallsignReserved
-from ..cert import sign_csr, revoke_pem, validate_reason, ReasonTypes, refresh_ocsp
+from ..cert.backend import sign_csr, revoke_pem, validate_reason, ReasonTypes, refresh_ocsp
 from ..productapihelpers import post_to_all_products
 from ..rmsettings import RMSettings
 from ..kchelpers import KCClient, KCUserData
