+++ src/rasenmaeher_api/productapihelpers.py
@@ -12,7 +12,7 @@ from libadvian.tasks import TaskMaster
 
 from .rmsettings import RMSettings
 from .mtlsinit import get_session_winit
-from .cert import refresh_ocsp
+from .cert.backend import refresh_ocsp
 
 LOGGER = logging.getLogger(__name__)
