+++ tests/test_db.py
@@ -34,7 +34,7 @@ from rasenmaeher_api.db.errors import (
 from rasenmaeher_api.jwtinit import jwt_init
 from rasenmaeher_api.mtlsinit import mtls_init
 from rasenmaeher_api.rmsettings import switchme_to_singleton_call, RMSettings
-from rasenmaeher_api.cert import get_crl
+from rasenmaeher_api.cert.backend import get_crl
 
 
 LOGGER = logging.getLogger(__name__)
